проверка на соблюдение критериям и техю заданию
