+++ js/main.js
@@ -1,14 +1,13 @@
 import { showAlertMessage } from './util.js';
-// import { getObjects } from './data.js';
 import { renderCards } from './popup.js';
-import { disableForm, disableFilters, activateFilters } from './form.js';
+import { disableForm, disableFilters, activateFilters } from './activation.js';
 import './validate-form-fields.js';
-import { resetForm, setOnFormSubmit } from './validate-form.js';
+import { resetForm, setOnFormSubmit } from './form.js';
 import { initMap, resetMap } from './map.js';
 import './form-fields.js';
 import { getData, sendData } from './api.js';
 import { showSuccessMessage, showErrorMessage } from './submit-message.js';
-import { initializeOfferFilters } from './map-filters.js';
+import { initializeOfferFilters, resetFilters } from './map-filters.js';
 
 setOnFormSubmit (async (data) => {
   try {
@@ -16,6 +15,7 @@ setOnFormSubmit (async (data) => {
     showSuccessMessage();
     resetForm();
     resetMap();
+    resetFilters();
   } catch {
     showErrorMessage();
   }
